Collapse the wire-step union and decompose sidecar wire contracts
Three coupled cleanups to the deploy-frame contracts in @intx/types:

- The `WorkflowStep` schema was ten near-identical arms whose
  per-variant fields were all optional passthrough, so the union
  validated almost nothing beyond the `kind` discriminant while adding
  ~110 lines of ceremony. Collapse it to one schema over the ten-kind
  enum plus the common fields. The load-bearing membership check is
  preserved; the handful of per-arm scalar validations that were
  enforced on some arms are intentionally dropped (the wire defers
  primitive-field validation to @intx/workflow).

- `AgentDeployWorkflow` restated the `definition` + `sources` +
  `approvedWireHash` field set and the byte-identical
  stepOrder-covered-by-sources narrow that
  `WorkflowProjectionWithSources` already carried -- a duplication this
  branch had widened. Express it as
  `WorkflowProjectionWithSources.and(...extras)` so the shared base and
  its narrow are defined once.

- With the union collapsed, sidecar.ts was still over 1000 lines.
  Extract the wire-step schema and the projection contracts to a new
  `wire-workflow.ts`, re-exporting the public names from sidecar.ts so
  `@intx/types/sidecar` consumers are unaffected. sidecar.ts drops to
  918 lines.
